finished pagination tests
