+++ tests/tensor_search/test_search.py
@@ -911,101 +911,6 @@ class TestVectorSearch(MarqoTestCase):
 
                 assert run()
 
-    def test_pagination_single_field(self):
-        vocab_source = "https://www.mit.edu/~ecprice/wordlist.10000"
-
-        vocab = requests.get(vocab_source).text.splitlines()
-        num_docs = 2000
-        
-        add_docs_caller(
-            config=self.config, index_name=self.index_name_1,
-            docs=[{"Title": "a " + (" ".join(random.choices(population=vocab, k=25))),
-                    "_id": str(i)
-                    }
-                  for i in range(num_docs)], auto_refresh=False
-        )
-        tensor_search.refresh_index(config=self.config, index_name=self.index_name_1)
-
-        for search_method in (SearchMethod.LEXICAL, SearchMethod.TENSOR):
-            for doc_count in [2000]:
-                # Query full results
-                full_search_results = tensor_search.search(
-                                        search_method=search_method,
-                                        config=self.config,
-                                        index_name=self.index_name_1,
-                                        text='a', 
-                                        result_count=doc_count)
-
-                for page_size in [5, 10, 100, 1000, 2000]:
-                    paginated_search_results = {"hits": []}
-
-                    for page_num in range(math.ceil(num_docs / page_size)):
-                        lim = page_size
-                        off = page_num * page_size
-                        page_res = tensor_search.search(
-                                        search_method=search_method,
-                                        config=self.config,
-                                        index_name=self.index_name_1,
-                                        text='a', 
-                                        result_count=lim, offset=off)
-                        
-                        paginated_search_results["hits"].extend(page_res["hits"])
-
-                    # Compare paginated to full results (length only for now)
-                    assert len(full_search_results["hits"]) == len(paginated_search_results["hits"])
-
-                    # TODO: re-add this assert when KNN incosistency bug is fixed
-                    # assert full_search_results["hits"] == paginated_search_results["hits"]
-                    
-    def test_pagination_break_limitations(self):
-        # Negative offset
-        for search_method in (SearchMethod.LEXICAL, SearchMethod.TENSOR):
-            for lim in [1, 10, 1000]:
-                for off in [-1, -10, -1000]:
-                    try:
-                        tensor_search.search(text=" ",
-                                            index_name=self.index_name_1, 
-                                            config=self.config, 
-                                            result_count=lim,
-                                            offset=off,
-                                            search_method=search_method)
-                        raise AssertionError
-                    except IllegalRequestedDocCount:
-                        pass
-        
-        # Negative limit
-        for search_method in (SearchMethod.LEXICAL, SearchMethod.TENSOR):
-            for lim in [0, -1, -10, -1000]:
-                for off in [1, 10, 1000]:
-                    try:
-                        tensor_search.search(text=" ",
-                                            index_name=self.index_name_1, 
-                                            config=self.config, 
-                                            result_count=lim,
-                                            offset=off,
-                                            search_method=search_method)
-                        raise AssertionError
-                    except IllegalRequestedDocCount:
-                        pass
-
-        # Going over 10,000 for offset + limit
-        mock_environ = {EnvVars.MARQO_MAX_RETRIEVABLE_DOCS: "10000"}
-        @mock.patch.dict(os.environ, {**os.environ, **mock_environ})
-        def run():
-            for search_method in (SearchMethod.LEXICAL, SearchMethod.TENSOR):
-                try:
-                    tensor_search.search(search_method=search_method,
-                                        config=self.config, index_name=self.index_name_1, text=' ', 
-                                        result_count=10000,
-                                        offset=1)
-                    raise AssertionError
-                except IllegalRequestedDocCount:
-                    pass
-            
-            return True
-
-        assert run()
-
     def test_empty_searchable_attributes(self):
         # Result should be empty whether paginated or not.
         docs = [
@@ -1033,60 +938,6 @@ class TestVectorSearch(MarqoTestCase):
             searchable_attributes=[], search_method="TENSOR"
         )
         assert res["hits"] == []
-    
-    def test_pagination_empty_searchable_attributes(self):
-        # Result should be empty whether paginated or not.
-        docs = [
-            {
-                "field_a": 0,
-                "field_b": 0, 
-                "field_c": 0
-            },
-            {
-                "field_a": 1,
-                "field_b": 1, 
-                "field_c": 1
-            }
-        ]
-
-        add_docs_caller(
-            config=self.config, index_name=self.index_name_1,
-            docs=docs, auto_refresh=False
-        )
-
-        tensor_search.refresh_index(config=self.config, index_name=self.index_name_1)
-
-        res = tensor_search.search(
-            config=self.config, index_name=self.index_name_1, text="some text",
-            searchable_attributes=[], search_method="TENSOR", offset=1
-        )
-        assert res["hits"] == []
-
-    def test_pagination_multi_field(self):
-        # Try pagination with 2 and 3 fields
-        docs = [
-            {
-                "field_a": 0,
-                "field_b": 0, 
-                "field_c": 0
-            },
-            {
-                "field_a": 1,
-                "field_b": 1, 
-                "field_c": 1
-            }
-        ]
-
-        add_docs_caller(
-            config=self.config, index_name=self.index_name_1,
-            docs=docs, auto_refresh=False
-        )
-
-        tensor_search.refresh_index(config=self.config, index_name=self.index_name_1)
-        for search_method in (SearchMethod.LEXICAL, SearchMethod.TENSOR):
-            # TODO: search with 2 and 3 searchable attribs, make sure result count is correct
-            # You can copy logic from test_pagination_single_field
-            pass
     
     def test_image_search_highlights(self):
         """does the URL get returned as the highlight? (it should - because no rerankers are being used)"""
